Adding `exports` key to be sorted above `main` (#60)
`package.json` supports `exports` maps, which will supersede `main` or `module` entries. Adding this to be sorted right above `main`. I chose to list it first, as node will prefer it over `main` and webpack will prefer it over `module` and typescript (when in resolution `nodenext`) will prefer it over types or typings.